+++ clients/fides-js/src/lib/consent-utils.ts
@@ -2,6 +2,7 @@ import { ConsentContext } from "./consent-context";
 import {
   ComponentType,
   ConsentMechanism,
+  CookieKeyConsent,
   EmptyExperience,
   FidesCookie,
   FidesOptions,
@@ -206,7 +207,8 @@ export const getTcfDefaultPreference = (tcfObject: TcfModelsRecord) =>
  */
 export const shouldResurfaceConsent = (
   experience: PrivacyExperience,
-  cookie: FidesCookie
+  cookie: FidesCookie,
+  savedConsent: CookieKeyConsent
 ): boolean => {
   if (experience.component === ComponentType.TCF_OVERLAY) {
     if (experience.meta?.version_hash) {
@@ -222,10 +224,15 @@ export const shouldResurfaceConsent = (
   ) {
     return false;
   }
-  // If not every notice has previous user consent, we need to resurface consent
+  // Always resurface if there is no prior consent
+  if (!savedConsent) {
+    return true;
+  }
+  // Lastly, if we do have a prior consent state, resurface if we find *any*
+  // notices that don't have prior consent in that state
   return Boolean(
     !experience.privacy_notices?.every((notice) =>
-      noticeHasConsentInCookie(notice, cookie.consent)
+      noticeHasConsentInCookie(notice, savedConsent)
     )
   );
 };