+++ CHANGELOG.md
@@ -18,10 +18,10 @@ The types of changes are:
 ## [Unreleased](https://github.com/ethyca/fides/compare/2.31.1...main)
 
 ### Changed
+- Removed location and regulation beta flag [#4660](https://github.com/ethyca/fides/pull/4660)
 - Modify `fides user login` to not store plaintext password in `~/.fides-credentials` [#4661](https://github.com/ethyca/fides/pull/4661)
 
 ### Fixed
-
 - Ignore 404 errors from Delighted and Kustomer when an erasure client is not found [#4593](https://github.com/ethyca/fides/pull/4593)
 
 ## [2.31.0](https://github.com/ethyca/fides/compare/2.30.1...2.31.0)
@@ -37,6 +37,7 @@ The types of changes are:
 
 ### Fixed
 - Fix issue where "x" button on Fides.js components overwrites saved preferences [#4649](https://github.com/ethyca/fides/pull/4649)
+- Initialize Fides.consent with default values from experience when saved consent cookie (fides_consent) does not exist [#4665](https://github.com/ethyca/fides/pull/4665)
 
 
 ## [2.30.1](https://github.com/ethyca/fides/compare/2.30.0...2.30.1)